requirements.txt
# Telegram Bot
python-telegram-bot==20.7
telegram==0.0.1

# REST API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Claude API
anthropic==0.18.1

# OpenAI (Whisper)
openai==1.12.0

# Local Whisper (опционально, для локального распознавания)
faster-whisper==1.0.0

# Ollama (для локальной LLM)
requests==2.31.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# Text-to-Speech
gTTS==2.5.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3

# Database
aiosqlite==0.19.0

# Logging
python-json-logger==2.0.7

# HTTP клиент (уже было, но явно указываем для Obsidian REST API и Ollama)
# requests==2.31.0  # уже добавлен выше
